+++ src/main/java/com/doubles/selfstudy/controller/QuestionBoardController.java
@@ -1,8 +1,6 @@
 package com.doubles.selfstudy.controller;
 
-import com.doubles.selfstudy.controller.request.QuestionBoardCommentRequest;
 import com.doubles.selfstudy.controller.request.QuestionBoardRequest;
-import com.doubles.selfstudy.controller.response.QuestionBoardCommentResponse;
 import com.doubles.selfstudy.controller.response.QuestionBoardResponse;
 import com.doubles.selfstudy.controller.response.Response;
 import com.doubles.selfstudy.dto.post.QuestionBoardDto;
@@ -91,55 +89,6 @@ public class QuestionBoardController {
             Authentication authentication
     ) {
         questionBoardService.deleteQuestionBoard(authentication.getName(), questionBoardId);
-
-        return Response.success();
-    }
-    
-    // 댓글 리스트
-    @GetMapping("/{questionBoardId}/comment")
-    public Response<Page<QuestionBoardCommentResponse>> questionBoardCommentList(
-            @PathVariable Long questionBoardId,
-            Pageable pageable,
-            Authentication authentication
-    ) {
-
-        return Response.success(
-                questionBoardService.questionBoardCommentList(questionBoardId, pageable)
-                        .map(QuestionBoardCommentResponse::fromQuestionBoardCommentDto)
-        );
-    }
-    
-    // 댓글 쓰기
-    @PostMapping("/{questionBoardId}/comment")
-    public Response<Void> createQuestionBoardComment(
-            @PathVariable Long questionBoardId,
-            @RequestBody QuestionBoardCommentRequest request,
-            Authentication authentication
-    ) {
-        questionBoardService.createQuestionBoardComment(
-                authentication.getName(), questionBoardId, request.getComment());
-
-        return Response.success();
-    }
-    
-    // 댓글 수정
-    @PutMapping("/{questionBoardId}/comment")
-    public Response<Void> updateQuestionBoardComment(
-            @PathVariable Long questionBoardId,
-            @RequestBody QuestionBoardCommentRequest request,
-            Authentication authentication
-    ) {
-
-
-        return Response.success();
-    }
-    
-    // 댓글 삭제
-    @DeleteMapping("/{questionBoardId}/comment")
-    public Response<Void> deleteQuestionBoardComment(
-
-    ) {
-
 
         return Response.success();
     }
