+++ src/test/java/com/doubles/selfstudy/controller/QuestionBoardControllerTest.java
@@ -1,6 +1,5 @@
 package com.doubles.selfstudy.controller;
 
-import com.doubles.selfstudy.controller.request.QuestionBoardCommentRequest;
 import com.doubles.selfstudy.controller.request.QuestionBoardRequest;
 import com.doubles.selfstudy.dto.post.QuestionBoardDto;
 import com.doubles.selfstudy.exception.DoubleSApplicationException;
@@ -343,55 +342,5 @@ class QuestionBoardControllerTest {
                 )
                 .andDo(print())
                 .andExpect(status().is(ErrorCode.ALREADY_LIKED.getStatus().value()));
-    }
-
-    @Test
-    @WithMockUser
-    void 질문_게시글_댓글_작성_성공() throws Exception {
-        // Given
-
-        // When
-
-        // Then
-        mockMvc.perform(post("/api/main/question_board/1/comment")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsBytes(new QuestionBoardCommentRequest("comment")))
-                )
-                .andDo(print())
-                .andExpect(status().isOk());
-    }
-
-    @Test
-    @WithAnonymousUser
-    void 질문_게시글_댓글_작성시_로그인하지_않은_경우_에러_발생() throws Exception {
-        // Given
-
-        // When
-
-        // Then
-        mockMvc.perform(post("/api/main/question_board/1/comment")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsBytes(new QuestionBoardCommentRequest("comment")))
-                )
-                .andDo(print())
-                .andExpect(status().is(ErrorCode.INVALID_TOKEN.getStatus().value()));
-    }
-
-    @Test
-    @WithMockUser
-    void 질문_게시글_댓글_작성시_게시물이_없는_경우_에러_반환() throws Exception {
-        // Given
-
-        // When
-        doThrow(new DoubleSApplicationException(ErrorCode.POST_NOT_FOUND))
-                .when(questionBoardService).createQuestionBoardComment(any(), any(), any());
-
-        // Then
-        mockMvc.perform(post("/api/main/question_board/1/comment")
-                        .contentType(MediaType.APPLICATION_JSON)
-                        .content(objectMapper.writeValueAsBytes(new QuestionBoardCommentRequest("comment")))
-                )
-                .andDo(print())
-                .andExpect(status().is(ErrorCode.POST_NOT_FOUND.getStatus().value()));
     }
 }
